requirements.txt
python-telegram-bot[all]==21.5
yt-dlp>=2024.10.7
aiosqlite>=0.20.0
psutil>=5.9.6
aiofiles>=23.2.1
python-dotenv>=1.0.1
aiohttp>=3.8.0
feedparser>=6.0.0
requests>=2.31.0
